# docs/Swagger.yaml
swagger: '2.0'
info:
  description: |
    Mushroom Cultivation System
  version: pre-alpha-1.0
  title: Mushroom++
paths:
...
  /hardware:
    post:
      summary: registers a new hardware
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - in: body
        name: Hardware
        required: true
        description: Hardware
        schema:
          $ref: '#/definitions/hardware'
      produces:
      - text/plain
      responses:
        200:
          description: Registered
        403:
          description: Unauthorized
        409:
          description: "Conflict, descriptive error message detailing if the hardware_id is taken."
          schema:
            example: "hardware_id \"bobs_test_hardware_1\" taken"
    get:
      summary: returns the list of hardwares associated with the user
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/hardware'
        403:
          description: Unauthorized
  /hardware/key/{hardware_key}:
    get:
      summary: returns the hardware with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: hardware_key
        in: path
        description: key of the Hardware
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/hardware'
        403:
          description: Unauthorized
        404:
          description: Not Found
    delete:
      summary: deletes the hardware with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: hardware_key
        in: path
        description: key of the Hardware
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        200:
          description: Deleted
        403:
          description: Unauthorized
        404:
          description: Not Found
    put:
      summary: updates the hardware with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: hardware_key
        in: path
        description: key of the Hardware
        required: true
        type: integer
      - in: body
        name: Hardware
        required: true
        description: Hardware
        schema:
          $ref: '#/definitions/hardware'
      produces:
      - text/plain
      responses:
        200:
          description: Updated
        403:
          description: Unauthorized
        404:
          description: Not Found
        409:
          description: "Conflict, descriptive error message detailing if the hardware_id is taken."
          schema:
            example: "hardware_id \"bobs_test_hardware_1\" taken"
  /hardware/key/{hardware_key}/sensor:
    get:
      summary: returns the current sensor data of the hardware associated with the given hardware key (polls realtime)
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: hardware_key
        in: path
        description: key of the Hardware 
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/sensor_entry'
        403:
          description: Unauthorized
        404:
          description: Not Found
  /specimen:
    post:
      summary: creates a new specimen
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - in: body
        name: Specimen
        required: true
        description: Specimen
        schema:
          $ref: '#/definitions/specimen'
      produces:
      - text/plain
      responses:
        200:
          description: Created
        403:
          description: Unauthorized
        409:
          description: "Conflict, descriptive error message detailing if the hardware_id is invalid, or the desired hardware_id already has an active specimen assigned, or if the specimen_type is invalid."
          schema:
            example: "hardware_id \"0123456789ABCDEF\" already has a specimen"
    get:
      summary: returns the list of active specimens associated with the user
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/specimen'
        403:
          description: Unauthorized
  /specimen/key/{specimen_key}:
    get:
      summary: returns the specimen with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: specimen_key
        in: path
        description: key of the Specimen
        required: true
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/specimen'
        403:
          description: Unauthorized (can be because of the Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
    delete:
      summary: deletes (marks as inactive, aka discards) the specimen with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: specimen_key
        in: path
        description: key of the Specimen
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        200:
          description: Deleted
        403:
          description: Unauthorized (can be because of the Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
    put:
      summary: updates the specimen with the given key
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: specimen_key
        in: path
        description: key of the Specimen
        required: true
        type: integer
      - in: body
        name: Specimen
        required: true
        description: Specimen
        schema:
          $ref: '#/definitions/specimen'
      produces:
      - text/plain
      responses:
        200:
          description: Updated
        403:
          description: Unauthorized (can be because of the Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
        409:
          description: "Conflict, descriptive error message detailing if the hardware_id is invalid, or the desired hardware_id already has an active specimen assigned, or if the specimen_type is invalid."
          schema:
            example: "hardware_id \"bobs_test_hardware_1\" not found"
  /specimen/key/{specimen_key}/sensor:
    get:
      summary: returns the sensor history associated with the given specimen key (can be filtered further with query strings)
      parameters:
      - in: header
        name: Authorization
        required: true
        type: string
        description: Bearer Authorization
      - name: specimen_key
        in: path
        description: key of the Specimen
        required: true
        type: integer
      - name: datetime_from
        in: query
        description: "Filter data based on a timespan.\nUnix datetime format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
        required: false
        type: integer
      - name: datetime_until
        in: query
        description: "Filter data based on a timespan.\nUnix datetime format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
        required: false
        type: integer
      produces:
      - application/json
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/sensor_entry'
        403:
          description: Unauthorized (can be because of the Specimen is marked as inactive, aka discarded, aka deleted)
        404:
          description: Not Found
...
definitions:
...
  hardware:
    type: object
    properties:
      hardware_key:
        description: "Nullable.\nReadonly value. (automatically generated on POST method)"
        type: integer
        example: 0
      hardware_id:
        description: "16 characters long HEX string, representing 64bit LoRaWAN id"
        type: string
        example: "0123456789ABCDEF"
      specimen_key:
        description: "Nullable\nReadonly value."
        type: integer
        example: 0
      desired_air_temperature:
        description: "Nullable\nReadonly value."
        type: number
        example: 21.7
      desired_air_humidity:
        description: "Nullable\nReadonly value."
        type: number
        example: 1.2
      desired_air_co2:
        description: "Nullable\nReadonly value."
        type: number
        example: 1.2
  specimen:
    type: object
    properties:
      specimen_key:
        description: "Nullable.\nReadonly value. (automatically generated on POST method)"
        type: integer
        example: 0
      planted_date:        
        description: "Readonly value.\nUnix datetime format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
        type: integer
        example: 1492992000000
      specimen_name:
        description: "User specified name."
        type: string
        example: "Champignon Test Cultivation Instance 1"
      specimen_type:
        description: "Type of mushroom (must be from the predefined list)."
        type: string
        example: "Champignon"
      specimen_description:
        description: "User specified description (notes)."
        type: string
        example: "Seeds are from AliExpress."
      desired_air_temperature:
        description: "Nullable, User specified value, also changes the assigned hardware's value."
        type: number
        example: 21.7
      desired_air_humidity:
        description: "Nullable, User specified value, also changes the assigned hardware's value."
        type: number
        example: 1.2
      desired_air_co2:
        description: "Nullable, User specified value, also changes the assigned hardware's value."
        type: number
        example: 1.2
      hardware_id:
        description: "Nullable.\nUser specified value, must be from the list of hardware ids associated with the user, or null."
        type: string
        example: "0123456789ABCDEF"
  sensor_entry:
    type: object
    properties:
      entry_key:
        description: "Nullabe (only used in history, not real time polling).\nReadonly value."
        type: integer
        example: 0
      entry_time:
        description: "Readonly value.\nUnix datetime format: number of milliseconds since the standard base time January 1, 1970, 00:00:00 GMT."
        type: integer
        example: 1492992000000
      air_temperature:
        description: "Readonly value."
        type: number
        example: 21.7
      air_humidity:
        description: "Readonly value."
        type: number
        example: 54.3
      air_co2:
        description: "Readonly value."
        type: number
        example: 1.2
      desired_air_temperature:
        description: "Nullable, Readonly value."
        type: number
        example: 21.7
      desired_air_humidity:
        description: "Nullable, Readonly value."
        type: number
        example: 1.2
      desired_air_co2:
        description: "Nullable, Readonly value."
        type: number
        example: 1.2
      ambient_air_temperature:
        description: "Readonly value."
        type: number
        example: 21.7
      ambient_air_humidity:
        description: "Readonly value."
        type: number
        example: 1.2
      ambient_air_co2:
        description: "Readonly value."
        type: number
        example: 1.2
      specimen_key:
        description: "Nullabe (only used in history, not real time polling).\nReadonly value.\nKey of the associated specimen."
        type: integer
        example: 145
...